Working tests with JUnit!
