+++ .travis.yml
@@ -10,6 +10,3 @@ before_install:
 
 git:
   depth: 3
-
-script:
-  - ./TeamCode/src/main/java/org/firstinspires/ftc/teamcode/rework/Tests/MathFunctionsTest.java
