+++ TeamCode/build.gradle
@@ -14,6 +14,5 @@
 // Include common definitions from above.
 apply from: '../build.common.gradle'
 dependencies {
-
-testImplementation 'org.junit.jupiter:junit-jupiter:5.4.2'
+    testImplementation 'junit:junit:4.12'
 }
